requirements.txt
requests==2.32.5
beautifulsoup4==4.12.3
playwright==1.49.1
pillow==10.1.0
pytesseract==0.3.10
